+++ classes/Gems/Default/AppointmentAction.php
@@ -218,7 +218,7 @@ class Gems_Default_AppointmentAction extends \Gems_Default_RespondentChildAction
      *
      * @return \Gems_Tracker_Respondent
      */
-    protected function getRespondent()
+    public function getRespondent()
     {
         if (! $this->_respondent) {
             $id = $this->_getParam(\Gems_Model::APPOINTMENT_ID);